refact: mock api entries separated into several files
